refacture: moved cli args into struct to reduce parameter count on modify
